polish: split lod_node.py in two files (one per class)
